Main: Hardcode the default network interface returned to lan0.
A temporary workaround for a problem: the default network interface string
is not accessible in TCP/IP stack driver memory for some machines. We get a trap
when trying to read it on one machine and get it successfully on another. So,
just fake the return result for the mean time.

git-svn-id: http://svn.netlabs.org/repos/vbox/trunk@112 8b531da7-8337-e511-8bd3-62de69f6b5d0